+++ src/Query/UpdateQueryBuilder.php
@@ -3,9 +3,6 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Worm\Query;
 
-use WoohooLabs\Larva\Query\Condition\ConditionBuilder;
-use WoohooLabs\Larva\Query\Condition\ConditionBuilderInterface;
-use WoohooLabs\Larva\Query\Select\SelectQueryBuilder as LarvaSelectQueryBuilder;
 use WoohooLabs\Larva\Query\Update\UpdateQueryBuilder as LarvaUpdateQueryBuilder;
 use WoohooLabs\Larva\Query\Update\UpdateQueryBuilderInterface;
 use WoohooLabs\Worm\Execution\QueryExecutor;
@@ -43,38 +40,16 @@ class UpdateQueryBuilder
         return $this;
     }
 
-    public function where(ConditionBuilderInterface $where): UpdateQueryBuilder
+    public function where(ConditionBuilder $where): UpdateQueryBuilder
     {
-        $this->queryBuilder->where($where);
+        $this->queryBuilder->where($where->getConditionBuilder());
 
         return $this;
     }
 
-    public function addWhereGroup(ConditionBuilderInterface $where): UpdateQueryBuilder
+    public function addWhereGroup(ConditionBuilder $where): UpdateQueryBuilder
     {
-        $this->queryBuilder->addWhereGroup($where);
-
-        return $this;
-    }
-
-    public function whereHas(
-        string $relationshipName,
-        ConditionBuilderInterface $conditionBuilder
-    ): UpdateQueryBuilder {
-        $relationship = $this->model->getRelationship($relationshipName);
-
-        $subselect = LarvaSelectQueryBuilder::create()
-            ->selectExpression("1")
-            ->from($relationship->getModel()->getTable())
-            ->where(
-                $conditionBuilder
-            );
-        $relationship->connectToParent($subselect);
-
-        $this->queryBuilder->addWhereGroup(
-            ConditionBuilder::create()
-                ->exists($subselect)
-        );
+        $this->queryBuilder->addWhereGroup($where->getConditionBuilder());
 
         return $this;
     }
@@ -95,11 +70,10 @@ class UpdateQueryBuilder
      */
     public function executeById($id): bool
     {
-        $this->queryBuilder
-            ->addWhereGroup(
-                ConditionBuilder::create()
-                    ->columnToValue($this->model->getPrimaryKey(), "=", $id)
-            );
+        $this->addWhereGroup(
+            ConditionBuilder::create()
+                ->columnToValue($this->model->getPrimaryKey(), "=", $id)
+        );
 
         return $this->queryExecutor->update($this->queryBuilder);
     }